# Optimize `check_indirect_parents`.