Currently we iterate all parents per level into a `Vec` via `calc_block_parents`, only to re-iterate them again in order to collect them into a `HashSet`. Not only does this incur one round of extra iterations, it also uses a std `HashSet`, which is sub-par compared to a `BlockHashSet`. This PR reduces one round of iterations by returning a `BlockHashSet` directly from the `ParentsManager`.  It also skips the usage of a std `Hashset` on the block hash, as is currently the case, as this is a post pow check, the usage of a `BlockHashSet` should be justified.

**benches:**

although i found it hard to benchmark classically due to the intricate nature of the parent manager, i logged both the old check, and the new one, in the wild, on a running node:

sample in-the-wild benchmarking code, in `post_pow_check`:

```rust
        let old: core::time::Duration = black_box(self.check_indirect_parents_old(ctx, header)?);
        let new: core::time::Duration = black_box(self.check_indirect_parents_new(ctx, header)?);
        info!("check_indirect_parents: winner: {:?}, old: {:?}, new: {:?}, times faster: {:?}", { if old.as_nanos() > new.as_nanos() {"new"} else {"old"} }, old, new, {if old > new { old.as_nanos() as f64 / new.as_nanos() as f64  } else { new.as_nanos() as f64 / old.as_nanos() as f64 }});
```

this is an excerpt of the results obtained:

```
2024-08-29 19:11:00.568+02:00 [INFO ] check_indirect_parents: winner: "new", old: 45.781µs, new: 25.492µs, times faster: 1.7958967519221716
2024-08-29 19:11:00.570+02:00 [INFO ] check_indirect_parents: winner: "new", old: 26.623µs, new: 11.203µs, times faster: 2.37641703115237
2024-08-29 19:11:00.588+02:00 [INFO ] check_indirect_parents: winner: "new", old: 46.416µs, new: 23.457µs, times faster: 1.9787696636398517
2024-08-29 19:11:00.602+02:00 [INFO ] check_indirect_parents: winner: "new", old: 50.332µs, new: 25.56µs, times faster: 1.969170579029734
2024-08-29 19:11:00.606+02:00 [INFO ] check_indirect_parents: winner: "new", old: 36.018µs, new: 18.641µs, times faster: 1.9321924789442626
2024-08-29 19:11:00.611+02:00 [INFO ] check_indirect_parents: winner: "new", old: 34.255µs, new: 19.226µs, times faster: 1.7817018620617913
2024-08-29 19:11:00.617+02:00 [INFO ] check_indirect_parents: winner: "new", old: 51.882µs, new: 23.682µs, times faster: 2.1907778059285534
2024-08-29 19:11:00.617+02:00 [INFO ] check_indirect_parents: winner: "new", old: 58.729µs, new: 27.039µs, times faster: 2.1720107992159474
2024-08-29 19:11:00.617+02:00 [INFO ] check_indirect_parents: winner: "new", old: 39.349µs, new: 22.241µs, times faster: 1.7692100175351828
2024-08-29 19:11:00.622+02:00 [INFO ] check_indirect_parents: winner: "new", old: 36.365µs, new: 17.467µs, times faster: 2.0819259174443236
2024-08-29 19:11:00.626+02:00 [INFO ] check_indirect_parents: winner: "new", old: 24.2µs, new: 14.419µs, times faster: 1.6783410777446426
2024-08-29 19:11:00.626+02:00 [INFO ] check_indirect_parents: winner: "new", old: 28.262µs, new: 12.535µs, times faster: 2.2546469884323894
```

which seems like a Ca. ~2x improvement. 

**Note:** 

1) Maintenance-wise, and concerning code-footprint, this PR simply does a mostly copy and pasted implementation of `calc_block_parents`.. as an alternative we could also reuse the existing `calc_block_parents`, and return an enum as either a `Vec<Vec<Hash>>` or `Vec<BlockHashSet>` depending on some parameter. Not sure what is the better approach.  

2) If it were possible to return the parents in the exact same ordering as found in `headers.parents`, without too much overhead, then the fastest way to perform the comparison might be to simply compare the `Vec<Vec<Hash>>`, with each other, but not sure if this is easily possible.